Expanded class and imported Maxs module
